update cache clearing for comment actions (#185)
Update cache clearing behavior for comment actions. Allow site cache
clearing setting to clear the site cache if any comment type (e.g.
default comments or product reviews) is posted, updated, spammed, or
trashed instead of only when a new (approved) comment is posted. This
fixes the issue of where if a new comment was posted but had to be
approved first it would not clear the site cache. There is no way to
know if the comment is new after the status transitions, which is why
this new behavior was chosen to be introduced instead of only updating
`Cache_Enabler::on_transition_comment_status()` as that could occur on
old comments. This builds upon the improved behavior brought in PR #169.