update requirement checks
Update requirement checks to be more clear on what Cache Enabler
requires to function properly. Checking the WP_CACHE constant is only
done in the settings page because if this was done everywhere the
warning can be shown when the Cache Enabler plugin is activated (even
when the constant was set) until the page is refreshed.