add definable plugin directory constant (#195)
When the `advanced-cache.php` file is loaded, it was previously looking
for `WP_PLUGIN_DIR . '/cache-enabler'`, which would prevent page
caching from working if the plugin was not installed here for whatever
reason (e.g. MU plugin).

This PR enables the path to be adjusted by defining a
`CACHE_ENABLER_DIR` constant, like in a site's `wp-config.php` file.