update settings file deletion handling (#205)
Update settings file deletion handling to also delete the parent
directory of the settings directory if empty. Previously only
`wp-content/settings/cache-enabler` would be deleted if empty. This
means if `wp-content/settings` is empty it will now also be deleted.
This will prevent leaving an empty directory that was created by Cache
Enabler if the plugin is deactivated.