Enable the page cache to be disabled without turning off the plugin
In situations where site owners might need to troubleshoot some caching issues but don't want to completely disable Cache Enabler, it's helpful to have the ability to pause the cache.

This commit adds a new "Cache Status" section to the top of the plugin settings screen, with a single option: "Enable the site cache". When checked (default behavior), it's assumed that the site owner wants the site cache enabled, and the plugin will proceed with checks around a missing `advanced-cache.php` file, the `WP_CACHE` constant, etc. However, when the page cache is disabled, these warnings are suppressed and both the constant and drop-in are removed. Re-enabling the page cache will then re-run `Cache_Enabler_Disk::setup()`, which puts these items back into place.