add new cache clearing structure for term actions
This extends #129 to also support term updates. The term archive will be cleared from cache on term updates as well as parent and child archives and pages associated with the updated term. In addition to that, parent term archives will be cleared on post updates, since they often contain posts from child terms.