also clear parent term archives before term is updated
When terms are updated, the parent could change too. To address this scenario, we clear the term's previous parent's term archives in the edit_terms hook.

In addition to that, the function signatures of term related functions got unified to support both: a WP_Term instance or a combination of term ID and taxonomy slug.