add new method
Add `Cache_Enabler::clear_post_terms_archives_cache()` method instead
of `Cache_Enabler::clear_taxonomies_archives_cache_by_post_id()`. This
will allow both a post instance or post ID. It is one of the few
methods that will be coming soon. It is not checking if the taxonomy is
public because this will be done in
`Cache_Enabler::clear_term_archive_cache()`. Checking if the taxonomy
is public would reduce the number of terms obtained and then sent
through `Cache_Enabler::clear_term_archive_cache()`, but I do not think
it is enough to check it twice (if someone disagrees please let me
know).