update requirements check (#249)
Update checking the plugin requirements. Use dynamic directory name and
path values in the notices instead of static values. For example,
instead of `wp-content/plugins/cache-enabler` get that partial
directory path by subtracting the `ABSPATH` value from the
`CACHE_ENABLER_DIR` value (this will still be
`wp-content/plugins/cache-enabler` for default installations). This
will be useful if custom directories are used and will be needed for

Add permissions check for the settings directory. This means now both
the cache and settings directories will be checked for writability.