update directory object filter (#252)
Update directory object filter to allow wildcard matching for files as
well as directories. This updated behavior was created when working on
PR #247 but ended up not being required in the end result. Cache
Enabler currently only uses wildcard matching for directories but we
may need this type of behavior in the future. Furthermore, it completes
the method as we do allow both directory and file paths to be sent
through it.

This updates what was added in PR #245.