update getting cache file
Update getting the cache file by having the
`Cache_Enabler_Disk::get_cache_file()` method set a class property for
the cache file path that has been generated. This will prevent having
to build the file path again when a cached page is created after
failing to be delivered.

This updates what was added in PR #211.